+++ Elements/src/ModelExtensions.cs
@@ -1,0 +1,41 @@
+using System.Collections.Generic;
+using System.Linq;
+
+namespace Elements
+{
+    /// <summary>
+    /// Model extension methods.
+    /// </summary>
+    public static class ModelExtensions
+    {
+        /// <summary>
+        /// Get all elements of a certain type from a specific model name in a dictionary of models.
+        /// </summary>
+        /// <param name="models">Dictionary of models keyed by string.</param>
+        /// <param name="modelName">The name of the model.</param>
+        /// <typeparam name="T">The type of element we want to retrieve.</typeparam>
+        /// <returns></returns>
+        public static List<T> AllElementsOfType<T>(this Dictionary<string, Model> models, string modelName) where T : Element
+        {
+            var elements = new List<T>();
+            models.TryGetValue(modelName, out var model);
+            if (model != null)
+            {
+                elements.AddRange(model.AllElementsOfType<T>());
+            }
+            return elements;
+        }
+
+        /// <summary>
+        /// Get all proxies of a certain type from a specific model name in a dictionary of models.
+        /// </summary>
+        /// <param name="models">Dictionary of models keyed by string</param>
+        /// <param name="modelName">The name of the model</param>
+        /// <typeparam name="T">The type of element we want to retrieve</typeparam>
+        /// <returns></returns>
+        public static List<ElementProxy<T>> AllProxiesOfType<T>(this Dictionary<string, Model> models, string modelName) where T : Element
+        {
+            return models.AllElementsOfType<T>(modelName).Proxies(modelName).ToList();
+        }
+    }
+}