+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - `Line.BestFit(IList<Vector3> points)`
 - `Vector3Extensions.BestFitLine(this IList<Vector3> points)`
 - `Polygon.FromAlignedBoundingBox2d(IEnumerable<Vector3> points, Vector3 axis, double minSideSize = 0.1)`
+- `Elements.IModel`
 
 ### Changed